+++ CHANGELOG.md
@@ -9,7 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Training now only displays reg_loss for models that have some regularization penalty configured.
 - Reduced the amount of log messages produced by `allennlp.common.file_utils`.
 - Fixed a bug where `PretrainedTransformerEmbedder` parameters appeared to be trainable
   in the log output even when `train_parameters` was set to `False`.
@@ -26,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a regression with logging in the distributed setting. Only the main worker should write log output to the terminal.
 - Pinned the version of boto3 for package managers (e.g. poetry).
 - Fixed issue #4330 by updating the `tokenizers` dependency.
+- reg_loss is only now returned for models that have some regularization penalty configured.
 
 ### Added